Made stacks process dependent. Instead of 1 global 32 bit stack, and 1 global 64 bit stack, each process now has it's own 32 bit and 64 bit mode, with their own stack corresponding stack. Each process will always start in 32 bit mode.
This change was made so threads do not block each other when
editing the stack, which is almost every type of instruction.

(Core Thread: A thread that is run by the deepest layer of a program.)
For direct core thread to core thread communication, it is recommended to use the registry.

Alternatively the memory can be written to and read from.

Registries are also much better at converting numbers between 32 bit and 64 bit.